pyproject.toml
[tool.poetry]
name = "conkyberry"
version = "0.0.0"
description = "Web application developed with Python and ReactJS for monitoring Raspberry Pi 4 model B"
authors = ["Rubén Rodríguez <rubenrod.018@gmail.com>"]

[tool.poetry.dependencies]
flask = "^1.1.2"
flask-graphql = "^2.0.1"
flask-mongoengine = "^0.9.5"
graphene-mongo = "^0.2.13"
psutil = "^5.7.2"
python = "^3.7"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]
pytest = "^6.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
